requirements.txt
python-telegram-bot<12

# Utils repository #
-r ../utils/requirements.txt